extract abstract class, fix associativity, cleanup (WP-854)
review fixes and changes
