+++ inc/Smartling/ContentTypes/Elementor/Elements/Accordion.php
@@ -13,7 +13,7 @@ class Accordion extends Unknown {
         $return = [];
         $knownRepeaterContent = ['tab_title', 'tab_content'];
         foreach ($this->settings['tabs'] ?? [] as $index => $tab) {
-            $key = 'tabs/' . $tab['_id'] ?? $index;
+            $key = 'tabs/' . ($tab['_id'] ?? $index);
             foreach ($knownRepeaterContent as $field) {
                 if (array_key_exists($field, $tab)) {
                     $return[$key][$field] = $tab[$field];
